[feat] Add a class to handle NamedTuple in a dick-like manner
When we use dicts or tuples that has same names again and again,
it is not clear for newcomers which variables we should include
as requirements.
For this reason, I recommend to use either BaseDict or BaseNamedTuple.
We should define the minimal requirements in a head of each file
and then we will know which variables to include.
By this feature, we are allowed to use NamedTuple like a dict-manner.
For more details, see the doc-string.