[fix] Fix mypy issues and modify the test accordingly
Since the previous codes had the default shuffle = True and the indices
shuffle before splitting, the test cases for CV and Holdout did not match.
More specifically, when I bring back the followings, I could reproduce
the original outputs:
1. Bring back _get_indices in BaseDataset
2. Make the default value of self.shuffle in BaseDataset True
3. Input shuffle = True in KFold instead of using ShuffleSplit
These reproduce the original outputs.
Note that KFold(shuffle=True) and ShuffleSplit() are not identical
and even when we input the same random_state, the results do not reproduce.